CORE-302 Move test to separate file
